Build `mountinfo` on multiple platforms
It functions the same as before this commit on Linux OSs that use the `sysfs` pseudo filesystem, and will now work on macOS through the use of the OS tools `stat` and `diskutil`. Windows, Solaris, Plan9, etc... are currently not supported. `go build` should work on those operating systems, and  this commit includes a shim for Windows: `device_windows.go`. I did notice that `GOOS=plan9 GOARCH=amd64 go build` failed because a transitive dependency does not support Plan9.

move functions into new platform-specific files. The entrypoint for the platform-specific code is the `discoverDeviceName` function.

move the `Test_DeviceName_Snapshots` function from `info_test.go` to `info_linux_test.go` so that the Linux test file will run a smoke test on `CWD` and also a test representing other mounted filesystems. Copy `Test_DeviceName_SmokeTest` from `info_linux_test.go` to `info_test.go` so that other operating systems will run the smoke test.

upgrade all dependencies and add a new dependency: `github.com/moby/sys/mountinfo` to handle reading mount info in order to find the `sysfs` mount point.

`discoverDeviceName` implementation for non-Unix/Linux/Windows operating systems. Currently returns a "not implemented" error.

`discoverDeviceName` implementation for macOS. Uses `stat` and `diskutil`.

`discoverDeviceName` implementation for Linux. Uses device identifier numbers and the `sysfs` pseudo filesystem to find block devices. Does not work for non-block devices (like Docker bind mounts).

functions that work on both Unix and Linux (and derivative) operating systems, but not on others. This is where the `golang.org/x/sys/unix` package is used to stat files in order to find the device identifier numbers.

`discoverDeviceName` implementation for Windows. Currently returns a "not implemented" error.

Run `go test -v` on various platforms.

`go test -v` will run a smoke test using the current working directory. The result should be a PASS, and if you provide the verbose option (`-v`), you'll see which device name was found from the current working directory.

If you don't have Linux readily available, you can use Docker. From the project directory, run this command to launch a Go container and `go test` the project:
```
docker run --rm --volume $(pwd):/src --volume mountinfo_test:/test --volume ${GOPATH:-${HOME}/go}:/go --interactive --workdir /test golang bash <<EOF
cp -R /src/* .
go test -v
EOF
```
The result should be a PASS, and if you provide the verbose option (`-v`), you'll see which device names were found in the tests.

Afterward, you can clean up the `mountinfo_test` volume if you want
```
docker volume rm mountinfo_test
```

There is a stub for windows - `device_windows.go` - but it currently returns an error, which is by design, so `go test` on Windows will result in a FAIL with a "not implemented on windows" error message.

We may be able to make this work on windows using the `DeviceIoControl` API, which is exposed in the `golang.org/x/sys/windows` package, alng with `STORAGE_PROPERTY_ID` and `STORAGE_DEVICE_ID_DESCRIPTOR` to get the SCSI device IDs. These device IDs may not be exactly what we need - `mountinfo` provides metrics for Prometheus, so the SCSI device IDs may be too long to be useful.

There is a catch-all stub for unsupported operating systems, such as Plan9 and Solaris - `device.go`. It will return a "not implemented on [GOOS]" error message, so `go test` will FAIL. This is by design.
