+++ device_windows.go
@@ -1,0 +1,12 @@
+package mountinfo
+
+import (
+	"fmt"
+	"runtime"
+
+	sglog "github.com/sourcegraph/log"
+)
+
+func discoverDeviceName(logger sglog.Logger, filePath string) (string, error) {
+	return "", fmt.Errorf("not implemented on %s", runtime.GOOS)
+}
